Prompt and response fully functioning
