+++ src/index.html
@@ -12,9 +12,13 @@
     <main class="main">
       <section class="promptsAndResponses">
         <div class="promptDiv">
-          <span class="prompt">guest@shasherazi ~ $ </span>
-          <input type="text" class="prompt-input" autofocus />
-          <div class="response">Dummy response</div>
+          <span class="prompt">guest@shasherazi ~ $</span>
+          <input type="text" class="prompt-input" id="input-0" disabled />
+        </div>
+        <div class="response">Dummy response</div>
+        <div class="promptDiv">
+          <span class="prompt">guest@shasherazi ~ $</span>
+          <input type="text" class="prompt-input" id="input-1" autofocus />
         </div>
       </section>
     </main>
